requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
user-agents>=2.2.0
pandas>=2.0.0
